# backend/docs/swagger.yaml
basePath: /api
...
host: localhost:8375
info:
  contact:
    email: support@sanctum.dev
    name: API Support
  description: Social media platform API with posts, comments, messaging, and friends
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Sanctum API
  version: "1.0"
paths:
...
  /ws/ticket:
    post:
      consumes:
      - application/json
      description: Generates a short-lived (60s), single-use ticket for WebSocket
        authentication to avoid sending JWTs in query parameters.
      produces:
      - application/json
      responses:
        "200":
          description: 'ticket: string, expires_in: int'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'error: unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: 'error: Redis not available'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Issue WebSocket ticket
      tags:
      - websocket
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"